Revert "Merge branch 'refs/heads/telnet-reconnect' into feature-graph-info-print"
This reverts commit 8691535a438669966db9f99cb044ff4f9210cfbb, reversing
changes made to 283cae012eea0bece31af47062f5647ea7874d23.

+++ Notifications/NotificationController.m
@@ -375,7 +375,7 @@
 	}
 
 	NSInteger sensorNum = [[sensorPopup selectedItem] tag];
-	NSAssert1((sensorNum < [sensorList count]), @"SensorPopup tag out of range: %ld", (long)sensorNum);
+	NSAssert1((sensorNum < [sensorList count]), @"SensorPopup tag out of range: %ld", sensorNum);
 	
 	NSMutableDictionary *sensor = [sensorList objectAtIndex: [[sensorPopup selectedItem] tag]];
